refactor(api): :art: move utils directory to common
